récupération du main

+++ src/main/java/tp/projetpappl/controllers/EnseignementController.java
@@ -34,6 +34,7 @@ import tp.projetpappl.items.TypeLecon;
 import tp.projetpappl.repositories.ContientRepository;
 import tp.projetpappl.repositories.EnseignantRepository;
 import tp.projetpappl.repositories.EnseignementRepository;
+import tp.projetpappl.items.Connection;
 import tp.projetpappl.repositories.SalleRepository;
 import tp.projetpappl.repositories.TypeLeconRepository;
 import tp.projetpappl.items.Connection;
@@ -47,7 +48,6 @@ public class EnseignementController {
 
     @Autowired
     private AuthHelper authHelper;
-
 
     @Autowired
     private EnseignementRepository enseignementRepository;
@@ -77,7 +77,6 @@ public class EnseignementController {
         List<Salle> salleList = salleRepository.findAll();
         returned.addObject("sallesList", salleList);
 
-        returned.addObject("user", user);
 
         returned.addObject("user", user);
 
@@ -102,13 +101,13 @@ public class EnseignementController {
         returned.addObject("enseignementsList", myList);
         returned.addObject("enseignantsList", Enseignants);
 
-        returned.addObject("user", user);
+
 
         returned.addObject("user", user);
 
         return returned;
     }
-    
+
     @RequestMapping(value = "editenseignement.do", method = RequestMethod.POST)
     public ModelAndView handleEditEnseignementPost(HttpServletRequest request) {
 
@@ -118,9 +117,19 @@ public class EnseignementController {
         }
 
 
+        String acronyme = request.getParameter("Acronyme");
+        if (acronyme != null) {
+            Enseignement enseignement = enseignementRepository.getByAcronyme(acronyme);
+            if (!authHelper.canModifyFiliere(request, enseignement.getFiliere())) {
+                ModelAndView forbidden = new ModelAndView("403");
+                forbidden.addObject("user", user);
+                return forbidden;
+            }
+        }
+
         ModelAndView returned;
 
-        String acronyme = request.getParameter("Acronyme");
+        //String acronyme = request.getParameter("Acronyme");
         if (acronyme != null) {
             //ID may exist
             Enseignement enseignement = enseignementRepository.getByAcronyme(acronyme);
@@ -149,14 +158,21 @@ public class EnseignementController {
             return new ModelAndView("redirect:login.do");
         }
 
+        String filiere = request.getParameter("Filiere");
+        if (!authHelper.canModifyFiliere(request, filiere)) {
+            ModelAndView forbidden = new ModelAndView("403");
+            forbidden.addObject("user", user);
+            return forbidden;
+        }
+
         ModelAndView returned;
 
         // Create or update enseignements
         String acronyme = request.getParameter("Acronyme");
         String nomEnseignement = request.getParameter("NomEnseignement");
-        String filiere = request.getParameter("Filiere");
+        //String filiere = request.getParameter("Filiere");
         String responsableStr = request.getParameter("InitialesEnseignant");
-        
+
         ArrayList<Contient> contientList = new ArrayList<Contient>();
         HashMap<String, String> nomEnseignants
                 = Tools.getArrayFromRequest(request, "e");
@@ -184,7 +200,7 @@ public class EnseignementController {
             BigInteger h = BigInteger.valueOf(heures);
             Contient contient = contientRepository.create(acronyme, typeLecon.getIntitule(), h, salle);
         }
-        
+
         for (TypeLecon typeLecon : typeLeconRepository.findAll()) {
             String heuresStr = request.getParameter(typeLecon.getIntitule());
             int heures = Tools.getIntFromString(heuresStr);
@@ -193,7 +209,7 @@ public class EnseignementController {
             BigInteger h = BigInteger.valueOf(heures);
             Contient contient = contientRepository.create(acronyme, typeLecon.getIntitule(), h, salle);
         }
-        
+
         returned = new ModelAndView("enseignement");
         returned.addObject("newenseignement", succes);
         returned.addObject("enseignantsList", enseignantRepository.findAll());
@@ -215,12 +231,20 @@ public class EnseignementController {
             return new ModelAndView("redirect:login.do");
         }
 
+        String acronyme = request.getParameter("Acronyme");
+        Enseignement enseignement = enseignementRepository.getByAcronyme(acronyme);
+        if (!authHelper.canModifyFiliere(request, enseignement.getFiliere())) {
+            ModelAndView forbidden = new ModelAndView("403");
+            forbidden.addObject("user", user);
+            return forbidden;
+        }
+
 
 
         ModelAndView returned;
 
         // Create or update enseignements
-        String acronyme = request.getParameter("Acronyme");
+        //String acronyme = request.getParameter("Acronyme");
 
         enseignementRepository.remove(acronyme);
 
@@ -231,7 +255,6 @@ public class EnseignementController {
         Collection<Enseignant> Enseignants = enseignantRepository.findAll();
         returned.addObject("enseignantsList", Enseignants);
 
-        returned.addObject("user", user);
 
         returned.addObject("user", user);
